% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalTraitsFinder.R
\name{optimalTraitsFinder}
\alias{optimalTraitsFinder}
\title{Optimal Traits Finder Class}
\description{
Finds the optimal no. traits for a phylogeny
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-optimalTraitsFinder-initialize}{\code{optimalTraitsFinder$new()}}
    \item \href{#method-optimalTraitsFinder-calculate_RF_distances}{\code{optimalTraitsFinder$calculate_RF_distances()}}
    \item \href{#method-optimalTraitsFinder-generate_RF_heatmap}{\code{optimalTraitsFinder$generate_RF_heatmap()}}
    \item \href{#method-optimalTraitsFinder-clone}{\code{optimalTraitsFinder$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-initialize}{}}}
\subsection{\code{optimalTraitsFinder$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$new(tree_dir)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-calculate_RF_distances"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-calculate_RF_distances}{}}}
\subsection{\code{optimalTraitsFinder$calculate_RF_distances()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$calculate_RF_distances()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-generate_RF_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-generate_RF_heatmap}{}}}
\subsection{\code{optimalTraitsFinder$generate_RF_heatmap()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$generate_RF_heatmap()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-clone}{}}}
\subsection{\code{optimalTraitsFinder$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
