% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertreeUtils.R
\name{supertreeUtils}
\alias{supertreeUtils}
\title{Supertree Utilities Class}
\description{
Collection of utilities for building a supertree
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{constituent_trees}}{Named list of trees making up supertree as phylo objects}

    \item{\code{constituent_tree_files}}{Named list of trees making up supertree as beast objects}

    \item{\code{mean_posteriors}}{Dataframe of mean posterior probabilities of constituent trees}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-supertreeUtils-initialize}{\code{supertreeUtils$new()}}
    \item \href{#method-supertreeUtils-mean_posterior_support}{\code{supertreeUtils$mean_posterior_support()}}
    \item \href{#method-supertreeUtils-clone}{\code{supertreeUtils$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-supertreeUtils-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-supertreeUtils-initialize}{}}}
\subsection{\code{supertreeUtils$new()}}{
  Initialises supertreeUtils Class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{supertreeUtils$new(tree_dir)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tree_dir}}{Location of consituent trees of supertree}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    constituent_trees

constituent_tree_files
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-supertreeUtils-mean_posterior_support"></a>}}
\if{latex}{\out{\hypertarget{method-supertreeUtils-mean_posterior_support}{}}}
\subsection{\code{supertreeUtils$mean_posterior_support()}}{
  Calculates mean posterior probability of nodes in each tree
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{supertreeUtils$mean_posterior_support()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    mean_posteriors
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-supertreeUtils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-supertreeUtils-clone}{}}}
\subsection{\code{supertreeUtils$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{supertreeUtils$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
