% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeProcessor.R
\name{treeProcessor}
\alias{treeProcessor}
\title{Tree Processor Class}
\description{
Read and plots trees
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-treeProcessor-initialize}{\code{treeProcessor$new()}}
    \item \href{#method-treeProcessor-tree_with_images}{\code{treeProcessor$tree_with_images()}}
    \item \href{#method-treeProcessor-clone}{\code{treeProcessor$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-treeProcessor-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-treeProcessor-initialize}{}}}
\subsection{\code{treeProcessor$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{treeProcessor$new(filename)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-treeProcessor-tree_with_images"></a>}}
\if{latex}{\out{\hypertarget{method-treeProcessor-tree_with_images}{}}}
\subsection{\code{treeProcessor$tree_with_images()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{treeProcessor$tree_with_images(image_dir)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-treeProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-treeProcessor-clone}{}}}
\subsection{\code{treeProcessor$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{treeProcessor$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
