% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalTraitsFinder.R
\name{optimalTraitsFinder}
\alias{optimalTraitsFinder}
\title{Optimal Traits Finder Class}
\description{
Finds the optimal no. traits for a phylogeny
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{trees}}{Named list of trees as phylo objects}

    \item{\code{tree_files}}{Named list of trees as beast objects}

    \item{\code{RF_distances}}{Dataframe of pairwise RF distances}

    \item{\code{mean_posteriors}}{Dataframe of mean posterior probabilities for each tree}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-optimalTraitsFinder-initialize}{\code{optimalTraitsFinder$new()}}
    \item \href{#method-optimalTraitsFinder-calculate_RF_distances}{\code{optimalTraitsFinder$calculate_RF_distances()}}
    \item \href{#method-optimalTraitsFinder-generate_RF_heatmap}{\code{optimalTraitsFinder$generate_RF_heatmap()}}
    \item \href{#method-optimalTraitsFinder-mean_posterior_support}{\code{optimalTraitsFinder$mean_posterior_support()}}
    \item \href{#method-optimalTraitsFinder-clone}{\code{optimalTraitsFinder$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-initialize}{}}}
\subsection{\code{optimalTraitsFinder$new()}}{
  Initialises optimalTraitsFinder class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$new(tree_dir)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{tree_dir}}{Location of tree files}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    trees

tree_files
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-calculate_RF_distances"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-calculate_RF_distances}{}}}
\subsection{\code{optimalTraitsFinder$calculate_RF_distances()}}{
  Calculates pairwise normalised RF distances for input trees
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$calculate_RF_distances()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    RF_distances
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-generate_RF_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-generate_RF_heatmap}{}}}
\subsection{\code{optimalTraitsFinder$generate_RF_heatmap()}}{
  Plots a heatmap of RF distances
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$generate_RF_heatmap()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Heatmap of RF distances
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-mean_posterior_support"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-mean_posterior_support}{}}}
\subsection{\code{optimalTraitsFinder$mean_posterior_support()}}{
  Calculates mean posterior probability of nodes in each tree
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$mean_posterior_support()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    mean_posteriors
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimalTraitsFinder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-optimalTraitsFinder-clone}{}}}
\subsection{\code{optimalTraitsFinder$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{optimalTraitsFinder$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
