% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potDataProcessor.R
\name{potDataProcessor}
\alias{potDataProcessor}
\title{Pot Data Processor Class}
\description{
Process pot data and generate basic plots
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-potDataProcessor-initialize}{\code{potDataProcessor$new()}}
    \item \href{#method-potDataProcessor-generate_revbayes_data}{\code{potDataProcessor$generate_revbayes_data()}}
    \item \href{#method-potDataProcessor-PC_plot}{\code{potDataProcessor$PC_plot()}}
    \item \href{#method-potDataProcessor-dendrogram}{\code{potDataProcessor$dendrogram()}}
    \item \href{#method-potDataProcessor-generate_pot_image}{\code{potDataProcessor$generate_pot_image()}}
    \item \href{#method-potDataProcessor-generate_pot_images}{\code{potDataProcessor$generate_pot_images()}}
    \item \href{#method-potDataProcessor-summarise_pot_profiles}{\code{potDataProcessor$summarise_pot_profiles()}}
    \item \href{#method-potDataProcessor-scale_taxa_data}{\code{potDataProcessor$scale_taxa_data()}}
    \item \href{#method-potDataProcessor-clone}{\code{potDataProcessor$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-initialize}{}}}
\subsection{\code{potDataProcessor$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$new(data)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-generate_revbayes_data"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-generate_revbayes_data}{}}}
\subsection{\code{potDataProcessor$generate_revbayes_data()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$generate_revbayes_data(
  save_dir = "",
  character_data_filename = "character_data",
  taxa_data_filename = "taxa_data"
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-PC_plot"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-PC_plot}{}}}
\subsection{\code{potDataProcessor$PC_plot()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$PC_plot(cols = c("PC1", "PC2"), add_elipses = FALSE)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-dendrogram"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-dendrogram}{}}}
\subsection{\code{potDataProcessor$dendrogram()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$dendrogram(n = 10)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-generate_pot_image"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-generate_pot_image}{}}}
\subsection{\code{potDataProcessor$generate_pot_image()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$generate_pot_image(data, name, save_dir = "")}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-generate_pot_images"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-generate_pot_images}{}}}
\subsection{\code{potDataProcessor$generate_pot_images()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$generate_pot_images(save_dir)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-summarise_pot_profiles"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-summarise_pot_profiles}{}}}
\subsection{\code{potDataProcessor$summarise_pot_profiles()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$summarise_pot_profiles()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-scale_taxa_data"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-scale_taxa_data}{}}}
\subsection{\code{potDataProcessor$scale_taxa_data()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$scale_taxa_data(save_dir = "", filename = "scaled_taxa_data")}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-potDataProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-potDataProcessor-clone}{}}}
\subsection{\code{potDataProcessor$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{potDataProcessor$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
